# gateway/internal/gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 多平台音乐网关 API
  version: 0.5.3
  description: |
    **多平台音乐网关 API** 是一组用于聚合 QQ 音乐 / 网易云音乐 / 酷狗音乐等多家音乐源的统一 RESTful 接口。
    网关在内部完成账号池调度、限流配额、计费扣减、Extra Usage 超额计费、余额扣款等控制逻辑，
    上游应用只需要通过统一的 API 接入即可。

    ## 协议总览
    - **基础路径**：所有业务接口位于 `/v1/...` 命名空间下。
    - **认证方式**：
      - **Cookie 会话**（前端控制台用）：登录后服务器下发 `sid` 会话 Cookie。
      - **API Key**（程序化调用用）：在 `Authorization: Bearer <api_key>` 请求头中携带。
    - **响应封装**：网关接口绝大多数返回 `{ code, message, data }` 包装结构（`code = 0` 表示成功），
      详情见下方 `Response` schema 定义。
    - **CSRF 保护**：使用 Cookie 会话时，所有非 GET/HEAD/OPTIONS 请求都需要在请求头中携带
      `X-CSRF-Token`（值与 Cookie `csrf_token` 一致）。API Key 模式不受 CSRF 限制。
    - **限流模型**：
      - **QPS**：按用户套餐限制每秒请求数。
      - **每日 / 每月配额**：免费版默认 100 次/日；超出后若开启 Extra Usage 则会按千次价从余额中扣费。

    ## 错误码
    全局错误码定义在 `ErrorCode` 字段：
    | code | 含义 |
    | ---- | ---- |
    | 0 | 成功 |
    | 400 | 请求参数错误 |
    | 401 | 未认证 / 凭据失效 |
    | 403 | 权限不足 |
    | 404 | 资源不存在 |
    | 409 | 资源冲突（重复创建、版本冲突等） |
    | 429 | 限流（QPS 或日/月配额耗尽） |
    | 500 | 服务器内部错误 |
  contact:
    name: 多平台音乐网关
    url: https://github.com/luoye663/qqmusicapi
servers:
  - url: /
    description: 当前实例
  - url: https://api.example.com
    description: 生产环境（占位，按部署实际域名替换）
...
security:
  - cookieAuth: []
  - apiKeyAuth: []
paths:
...
  /v1/auth/api-keys:
    get:
      tags: [API Key]
      summary: 列出当前用户的 API Key
      responses:
        "200":
          description: API Key 列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/APIKeyListEnvelope" }
    post:
      tags: [API Key]
      summary: 创建 API Key
      description: 创建后**只在本次响应中**返回完整 Key 明文，请妥善保存。新 Key 的套餐继承用户当前套餐。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAPIKeyRequest" }
      responses:
        "200":
          description: 创建成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateAPIKeyResponseEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/auth/api-keys/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
        description: API Key ID（不是 Key 明文）
    get:
      tags: [API Key]
      summary: 获取单个 API Key 元信息
      responses:
        "200":
          description: API Key 详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/APIKeyEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [API Key]
      summary: 更新 API Key 名称 / 备注 / 限流 / 过期时间
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateAPIKeyRequest" }
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/APIKeyEnvelope" }
    delete:
      tags: [API Key]
      summary: 注销 API Key
      description: 软删除，撤销后无法再恢复。
      responses:
        "200":
          description: 注销成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmptyEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/auth/api-keys/{id}/enabled:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    post:
      tags: [API Key]
      summary: 启用 / 停用 API Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        "200":
          description: 状态已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmptyEnvelope" }

  # ========== 套餐 ==========
...
  /v1/billing/extra-usage:
    get:
      tags: [Extra Usage]
      summary: 获取当前用户的 Extra Usage 配置
      responses:
        "200":
          description: 配置
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExtraUsageSettingsEnvelope" }
    put:
      tags: [Extra Usage]
      summary: 更新 Extra Usage 配置
      description: |
        - `enabled` 为 `true` 时，月度 / 每日配额耗尽后将按 `overagePricePer1k` 从余额中扣费。
        - `monthlyCapCents` 设置月度超额上限（防意外大额扣费），0 表示不设上限。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateExtraUsageRequest" }
      responses:
        "200":
          description: 已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExtraUsageSettingsEnvelope" }
...
  /v1/billing/extra-usage/charges:
    get:
      tags: [Extra Usage]
      summary: 列出超额扣费聚合（按月）
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 12, maximum: 60 } }
      responses:
        "200":
          description: 月度聚合列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OverageChargesEnvelope" }
  /v1/billing/extra-usage/events:
    get:
      tags: [Extra Usage]
      summary: 列出超额扣费事件明细
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 500 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: 事件明细
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OverageEventsEnvelope" }

  # ========== 用量 ==========
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
      description: 通过 `/v1/auth/login` 颁发的会话 Cookie，浏览器自动携带；同源调用无需手动配置。
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: "在 `Authorization: Bearer <api_key>` 中携带。"
...
  responses:
...
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
...
    Response:
      type: object
      required: [code, message]
      properties:
        code: { type: integer, example: 0, description: "0=成功；非 0 见错误码表" }
        message: { type: string, example: "ok" }
        data: { description: "实际业务数据，类型见各接口" }
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          example: { code: 400, message: "invalid body", data: null }
    EmptyEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data: { type: object, additionalProperties: true }
...
    APIKey:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        prefix: { type: string, description: 仅返回前缀，明文 Key 只在创建时给一次 }
        planId: { type: string }
        status: { type: string, enum: [active, disabled] }
        scopes: { type: array, items: { type: string } }
        ipAllowlist: { type: array, items: { type: string } }
        rateLimitRpm: { type: integer, nullable: true }
        rateLimitDaily: { type: integer, format: int64, nullable: true }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time, nullable: true }
        lastUsedAt: { type: string, format: date-time, nullable: true }
    APIKeyEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/APIKey" }
    APIKeyListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              properties:
                items: { type: array, items: { $ref: "#/components/schemas/APIKey" } }
                total: { type: integer }
    CreateAPIKeyRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 64 }
        description: { type: string }
        scopes: { type: array, items: { type: string } }
        ipAllowlist: { type: array, items: { type: string } }
        rateLimitRpm: { type: integer }
        rateLimitDaily: { type: integer, format: int64 }
        expiresAt: { type: string, format: date-time }
    CreateAPIKeyResponseEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              properties:
                key: { type: string, description: "完整明文 Key（只此一次）" }
                record: { $ref: "#/components/schemas/APIKey" }
    UpdateAPIKeyRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        scopes: { type: array, items: { type: string } }
        ipAllowlist: { type: array, items: { type: string } }
        rateLimitRpm: { type: integer }
        rateLimitDaily: { type: integer, format: int64 }
        expiresAt: { type: string, format: date-time, nullable: true }

    # ---- 套餐 / 订单 / 余额 / Extra Usage ----
...
    ExtraUsageSettings:
      type: object
      properties:
        userId: { type: string }
        enabled: { type: boolean }
        monthlyCapCents: { type: integer, format: int64, description: "0=不设上限" }
        currency: { type: string }
        updatedAt: { type: string, format: date-time }
    ExtraUsageSettingsEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/ExtraUsageSettings" }
    UpdateExtraUsageRequest:
      type: object
      properties:
        enabled: { type: boolean }
        monthlyCapCents: { type: integer, format: int64 }
...
    OverageCharge:
      type: object
      properties:
        userId: { type: string }
        month: { type: string }
        requests: { type: integer, format: int64 }
        amountCents: { type: integer, format: int64 }
    OverageChargesEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              properties:
                items: { type: array, items: { $ref: "#/components/schemas/OverageCharge" } }
    OverageEvent:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        ts: { type: string, format: date-time }
        provider: { type: string }
        endpoint: { type: string }
        amountCents: { type: integer, format: int64 }
        units: { type: integer, format: int64, description: "本次扣费单位数（默认 1）" }
        balanceAfterCents: { type: integer, format: int64 }
    OverageEventsEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              properties:
                items: { type: array, items: { $ref: "#/components/schemas/OverageEvent" } }
                total: { type: integer }
                limit: { type: integer }
                offset: { type: integer }

    # ---- 用量 ----
...